feat(codegen): load abi from environments
