+++ codegen/src/codegen/dispatcher.rs
@@ -1,13 +1,12 @@
 //! Code generator for EVM dispatcher.
 
-use std::collections::BTreeMap;
-
 use crate::{
     codegen::code::ExtFunc,
     wasm::{self, Env, Functions, ToLSBytes},
     Error, JumpTable, MacroAssembler, Result,
 };
-use wasmparser::{FuncType, Operator};
+use std::collections::BTreeMap;
+use wasmparser::FuncType;
 use zabi::Abi;
 
 /// Code generator for EVM dispatcher.
@@ -44,14 +43,13 @@ impl Dispatcher {
     /// Emit compiled code to the given buffer.
     pub fn finish(&mut self, selectors: Functions<'_>, table: &mut JumpTable) -> Result<Vec<u8>> {
         if selectors.is_empty() {
-            return Err(Error::SelectorNotFound);
+            return Ok(Default::default());
         }
 
         self.asm._push0()?;
         self.asm._calldataload()?;
         self.asm.push(&[0xe0])?;
         self.asm._shr()?;
-
         let mut len = selectors.len();
         for (_, func) in selectors.iter() {
             self.emit_selector(func, len == 1)?;
@@ -71,33 +69,6 @@ impl Dispatcher {
         }
 
         Err(Error::FuncNotImported(name.into()))
-    }
-
-    /// Load function ABI.
-    fn load_abi(&mut self, selector: &wasm::Function<'_>) -> Result<Abi> {
-        let mut reader = selector.body.get_operators_reader()?;
-
-        let Operator::I32Const { value: offset } = reader.read()? else {
-            return Err(Error::InvalidSelector);
-        };
-        let Operator::I32Const { value: length } = reader.read()? else {
-            return Err(Error::InvalidSelector);
-        };
-
-        // Validate zinkc helper `emit_abi`
-        let Operator::Call {
-            function_index: index,
-        } = reader.read()?
-        else {
-            return Err(Error::InvalidSelector);
-        };
-
-        if !self.env.imports.is_emit_abi(index) {
-            return Err(Error::FuncNotImported("emit_abi".into()));
-        }
-
-        let abi = self.env.data.load(offset, length as usize)?;
-        Abi::from_hex(String::from_utf8_lossy(&abi)).map_err(Into::into)
     }
 
     /// Emit return of ext function.
@@ -186,7 +157,7 @@ impl Dispatcher {
 
     /// Emit selector to buffer.
     fn emit_selector(&mut self, selector: &wasm::Function<'_>, last: bool) -> Result<()> {
-        let abi = self.load_abi(selector)?;
+        let abi = self.env.load_abi(selector)?;
 
         // TODO: refactor this. (#206)
         self.abi.push(abi.clone());
